refactor(product): Phase 7 cleanup, slice manifest, enforce boundaries
- Move InternalOutboxEndpoints to Infrastructure/Outbox; delete empty Endpoints/.
- Unskip 4 NetArchTest layout rules; promote LayoutAnalyzer severities to Error.
- Fix dead PRDLAY002 rule (FeatureSlice never fired due to trailing-dot prefix bug).
