+++ product-microservice/Product.Tests/Architecture/LayoutTests.cs
@@ -6,7 +6,7 @@ public class LayoutTests
 {
     private static readonly System.Reflection.Assembly ProductServiceAssembly = typeof(Program).Assembly;
 
-    [Fact(Skip = "Enabled in Phase 7")]
+    [Fact]
     public void Domain_DoesNotReference_InfrastructureOrFeatures()
     {
         var result = Types.InAssembly(ProductServiceAssembly)
@@ -21,7 +21,7 @@ public class LayoutTests
             + string.Join(", ", result.FailingTypeNames ?? []));
     }
 
-    [Fact(Skip = "Enabled in Phase 7")]
+    [Fact]
     public void Features_DoNotReference_OtherFeatureSlices()
     {
         var featureTypes = Types.InAssembly(ProductServiceAssembly)
@@ -66,7 +66,7 @@ public class LayoutTests
             "Features.<X> may not reference Features.<Y> for X != Y: " + string.Join(", ", offenders));
     }
 
-    [Fact(Skip = "Enabled in Phase 7")]
+    [Fact]
     public void Infrastructure_DoesNotReference_Features()
     {
         var result = Types.InAssembly(ProductServiceAssembly)
@@ -81,7 +81,7 @@ public class LayoutTests
             + string.Join(", ", result.FailingTypeNames ?? []));
     }
 
-    [Fact(Skip = "Enabled in Phase 7")]
+    [Fact]
     public void Contracts_DoNotReference_OtherProductServiceNamespaces()
     {
         var result = Types.InAssembly(ProductServiceAssembly)
